requirements.txt
absl-py==0.15.0
actionlib==1.13.2
addict==2.4.0
affine==2.3.0
aiohttp==3.8.1
aiosignal==1.2.0
alphashape==1.3.1
angles==1.9.13
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
bondpy==1.8.6
cachetools==5.1.0
camera-calibration==1.15.3
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
click-log==0.3.2
click-plugins==1.1.1
cligj==0.7.2
configparser==5.2.0
controller-manager==0.19.4
controller-manager-msgs==0.19.4
cv-bridge==1.15.0
cycler==0.11.0
Cython==0.29.30
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
descartes==1.1.0
diagnostic-analysis==1.10.3
diagnostic-common-diagnostics==1.10.3
diagnostic-updater==1.10.3
docker-pycreds==0.4.0
dynamic-reconfigure==1.7.1
entrypoints==0.4
executing==0.8.3
filelock==3.7.1
Fiona==1.8.21
Flask==2.1.1
flatbuffers==1.12
fonttools==4.29.1
freetype-py==2.2.0
frozenlist==1.3.0
gast==0.4.0
gazebo_plugins==2.9.1
gazebo_ros==2.9.1
GDAL==3.0.4
gdown==3.15.0
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.14
geopandas==0.10.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googledrivedownloader==0.4
gql==0.2.0
graphql-core==1.1
grpcio==1.34.1
h5py==3.1.0
hydra-core==0.11.3
idna==3.3
image-geometry==1.15.0
imageio==2.17.0
importlib-metadata==4.11.3
importlib-resources==5.4.0
interactive-markers==1.12.0
ipykernel==6.9.1
ipyleaflet==0.16.0
ipython==8.1.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
joint_state_publisher==1.15.0
joint_state_publisher_gui==1.15.0
json5==0.9.6
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-packaging==0.11.1
jupyter-server==1.13.5
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.2
keras==2.9.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
laser_geometry==1.6.7
laspy==2.1.2
libclang==14.0.1
llvmlite==0.33.0
Markdown==3.3.7
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
message-filters==1.15.9
mistune==0.8.4
mpmath==1.2.1
multidict==6.0.2
munch==2.5.0
nbclassic==0.3.5
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.7.1
node==1.0
nose==1.3.7
notebook==6.4.8
numba==0.50.1
numpy==1.23.0
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
odict==1.8.1
omegaconf==1.4.1
open3d==0.10.0.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
plumber==1.7
plyfile==0.7.4
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.2
pycodestyle==2.8.0
pycparser==2.21
pyglet==1.5.23
Pygments==2.11.2
pymesh==1.0.2
PyOpenGL==3.1.0
pyparsing==3.0.7
pyproj==3.3.1
pyquaternion==0.9.9
pyrender==0.1.45
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-louvain==0.16
python-qt-binding==0.4.3
pytorch-metric-learning==0.9.99
pytz==2021.3
pyvista==0.34.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
rasterio==1.2.10
rasterstats==0.16.0
rdflib==6.1.1
requests==2.27.1
requests-oauthlib==1.3.1
resource_retriever==1.12.6
rosbag==1.15.9
rosboost-cfg==1.15.7
rosclean==1.15.7
roscreate==1.15.7
rosgraph==1.15.9
roslaunch==1.15.9
roslib==1.15.7
roslint==0.12.0
roslz4==1.15.9
rosmake==1.15.7
rosmaster==1.15.9
rosmsg==1.15.9
rosnode==1.15.9
rosparam==1.15.9
rospy==1.15.9
rosservice==1.15.9
rostest==1.15.9
rostopic==1.15.9
rosunit==1.15.7
roswtf==1.15.9
rqt-moveit==0.5.9
rqt-reconfigure==0.5.3
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.13
rqt-rviz==0.6.1
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.11
rqt_dep==0.4.10
rqt_graph==0.4.14
rqt_gui==0.5.2
rqt_gui_py==0.5.2
rqt_image_view==0.4.16
rqt_launch==0.4.9
rqt_logger_level==0.4.11
rqt_msg==0.4.9
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.10
rqt_publisher==0.4.9
rqt_py_common==0.5.2
rqt_py_console==0.4.9
rqt_robot_steering==0.5.12
rqt_runtime_monitor==0.5.8
rqt_service_caller==0.4.9
rqt_shell==0.4.10
rqt_srv==0.4.8
rqt_tf_tree==0.6.2
rqt_top==0.4.9
rqt_topic==0.4.12
rqt_web==0.4.9
rsa==4.8
Rtree==0.9.7
rviz==1.14.5
scikit-image==0.16.2
scikit-learn==1.0.2
scipy==1.8.0
scooby==0.5.12
Send2Trash==1.8.0
sensor-msgs==1.13.1
sentry-sdk==1.6.0
Shapely==1.8.1.post1
shortuuid==1.0.9
simplejson==3.17.6
simplekml==1.3.6
six==1.15.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
smmap==5.0.0
sniffio==1.2.0
snuggs==1.4.7
stack-data==0.2.0
subprocess32==3.5.4
sympy==1.10.1
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.3
tensorflow-estimator==2.5.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.6.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
threadpoolctl==3.1.0
tifffile==2022.5.4
toml==0.10.2
tomlkit==0.10.0
topic-tools==1.15.9
torch @ https://open3d-downloads.b-cdn.net/torch-1.8.2-cp38-cp38-linux_x86_64.whl
torch-cluster==1.6.0
torch-geometric==1.7.2
torch-points-kernels==0.6.10
torch-points3d==1.3.0
torch-scatter==2.0.9
torch-sparse==0.6.14
torchfile==0.1.0
torchnet==0.0.4
torchvision==0.9.2+cu111
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
traittypes==0.2.1
trimesh==3.10.2
typing-extensions==3.7.4.3
urllib3==1.26.8
visdom==0.1.8.9
vtk==9.1.0
wandb==0.8.36
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.1.1
widgetsnbextension==3.5.2
wrapt==1.12.1
wslink==1.6.5
xacro==1.14.6
xyzservices==2022.4.0
yarl==1.7.2
zipp==3.7.0
zope.component==5.0.1
zope.deferredimport==4.4
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
zope.lifecycleevent==4.3
zope.proxy==4.5.0
